Calculate selection chrome coordinate transformation manually
For some reason apply_relative_transform_to_point didn't work when changing
workspace. (the _first_ workspace switch when going from only one minimap
visible to all visible)

Even when adjusting for destintationX of minimapActor.. Not really sure why, so
this isn't the most satisfying fix.

A manual transform would be needed if there's multiple actors in the ancestor
chain that is animating though. (ie. to account for destinationX instead of x)